Reverify and Recreate working. Also updated Delete Account

+++ homebook/src/screens/SideMenu.js
@@ -32,6 +32,14 @@ class SideMenu extends Component {
         }
     });
 
+    confirmInput = () => {
+        if (this.state.currentPassword == '') {
+            alert("Please input the current password");
+            return false;
+        }
+        return true;
+    };
+
     deleteAccount = () => {
         var user = firebase.auth().currentUser;
         var db = firebase.firestore();
@@ -41,35 +49,28 @@ class SideMenu extends Component {
             this.state.currentPassword
         );
 
-        user.reauthenticateAndRetrieveDataWithCredential(credential).then(() => {
-            // User re-authenticated.
-            db.collection("users").where("uid", "==", firebase.auth().currentUser.uid).get().then((querySnapshot) => {
-                querySnapshot.forEach((doc) => {
-                    db.collection("users").doc(doc.id).delete();
-                })
-            }).then(() => {
-                user.delete().then(() => {
-                    // User deleted.
-                    Navigation.popToRoot(this.props.componentId);
-                }).catch(() => {
-                    // An error happened.
+        if (this.confirmInput()) {
+            user.reauthenticateAndRetrieveDataWithCredential(credential).then(() => {
+                // User re-authenticated.
+                db.collection("users").where("uid", "==", firebase.auth().currentUser.uid).get().then((querySnapshot) => {
+                    querySnapshot.forEach((doc) => {
+                        db.collection("users").doc(doc.id).delete();
+                    })
+                }).then(() => {
+                    user.delete().then(() => {
+                        // User deleted.
+                        Navigation.popToRoot(this.props.componentId);
+                    }).catch(() => {
+                        // An error happened.
+                    });
+                }).catch(function (error) {
+                    alert("Error getting documents: " + error);
                 });
-            }).catch(function (error) {
-                alert("Error getting documents: " + error);
+            }).catch(() => {
+                // An error happened.
+                alert("Password Incorrect");
             });
-        }).catch(() => {
-            // An error happened.
-            alert("Password Incorrect");
-        });
-
-
-
-
-
-
-
-
-
+        }
     };
 
     closeSideMenu = () => Navigation.mergeOptions(this.props.componentId, {
